Replace vuex to props for child profile components
