+++ src/frontend/src/modules/profile/components/ProfileAddressItem.vue
@@ -6,7 +6,7 @@
         <button
           type="button"
           class="icon"
-          @click="editAddress()"
+          @click="$emit('edit')"
         >
           <span class="visually-hidden">Изменить адрес</span>
         </button>
@@ -18,13 +18,6 @@
 </template>
 
 <script>
-import { mapMutations } from "vuex";
-import {
-  START_EDIT_ADDRESS,
-  SET_PROFILE_ADDRESS,
-  SHOW_CREATE_NEW_ADDRESS_FORM,
-} from "@/store/mutation-types";
-
 export default {
   name: "ProfileAddressItem",
   props: {
@@ -34,30 +27,13 @@ export default {
       default: () => ({}),
     },
   },
-
   computed: {
     flat() {
       return this.address.flat.length > 0 ? `, кв. ${this.address.flat}` : "";
     },
-
     fullAddress() {
       return `${this.address.street}, д. ${this.address.building}${this.flat}`;
     },
-  },
-
-  methods: {
-    editAddress() {
-      this[SHOW_CREATE_NEW_ADDRESS_FORM](false);
-      this[SET_PROFILE_ADDRESS](this.address);
-      this.$emit("edit", this.address.id);
-      this[START_EDIT_ADDRESS](true);
-    },
-
-    ...mapMutations("address", [
-      START_EDIT_ADDRESS,
-      SET_PROFILE_ADDRESS,
-      SHOW_CREATE_NEW_ADDRESS_FORM,
-    ]),
   },
 };
 </script>
