+++ src/frontend/src/views/ProfilePage.vue
@@ -19,9 +19,19 @@
         :key="address.id"
         class="layout__address"
       >
-        <ProfileAddressController
+        <ProfileAddressForm
+          v-if="startedEditAddress"
+          delete-button
           :address="address"
-          :user-id="user.id"
+          :errors="errors"
+          @save="saveAddress()"
+          @delete="deleteSelectedAddress(address.id)"
+          @close="toggleNewAddressFormCreating(false)"
+        />
+        <ProfileAddressItem
+          v-else
+          :address="address"
+          @edit="editAddress(address.id)"
         />
       </div>
     </template>
@@ -31,10 +41,10 @@
       class="layout__address"
     >
       <ProfileAddressForm
-        :is-delete-visible="false"
         :address="profileAddress"
-        :user-id="user.id"
-        @close="closeAddressForm()"
+        :errors="errors"
+        @save="saveAddress()"
+        @close="toggleNewAddressForm(false)"
       />
     </div>
 
@@ -43,7 +53,7 @@
         border
         type="button"
         :disabled="!isNewAddressFormCreationAvailable"
-        @click="toggleNewAddressFormCreation()"
+        @click="toggleNewAddressForm(true)"
       >
         Добавить новый адрес
       </AppButton>
@@ -52,36 +62,115 @@
 </template>
 
 <script>
-import { mapState, mapGetters, mapMutations } from "vuex";
-import { SHOW_CREATE_NEW_ADDRESS_FORM } from "@/store/mutation-types";
+import { mapState, mapGetters, mapMutations, mapActions } from "vuex";
+import { validateForm } from "@/services/formValidation";
+import {
+  RESET_PROFILE_ADDRESS,
+  SET_PROFILE_ADDRESS,
+  SHOW_CREATE_NEW_ADDRESS_FORM,
+  START_EDIT_ADDRESS
+} from "@/store/mutation-types";
+
 import ProfileUser from "@/modules/profile/components/ProfileUser.vue";
-import ProfileAddressController from "@/modules/profile/components/ProfileAddressController.vue";
 import ProfileAddressForm from "@/modules/profile/components/ProfileAddressForm.vue";
+import ProfileAddressItem from "@/modules/profile/components/ProfileAddressItem.vue";
 
 export default {
   name: "ProfilePage",
   components: {
+    ProfileAddressItem,
     ProfileUser,
-    ProfileAddressController,
     ProfileAddressForm,
+  },
+  data() {
+    return {
+      errors: [],
+    };
   },
   computed: {
     ...mapState("address", [
       "addresses",
       "profileAddress",
       "isNewAddressFormCreationVisible",
+      "startedEditAddress",
     ]),
     ...mapState("auth", ["user"]),
     ...mapGetters("address", ["isNewAddressFormCreationAvailable"]),
   },
   methods: {
-    closeAddressForm() {
+    toggleNewAddressForm(state) {
+      this[SHOW_CREATE_NEW_ADDRESS_FORM](state);
+    },
+    closeEditAddressForm() {
+      this[START_EDIT_ADDRESS](false);
+      this[RESET_PROFILE_ADDRESS]();
+    },
+    async deleteSelectedAddress(id) {
+      await this.deleteAddress(id);
+      await this.fetchAddresses();
+      this.closeEditAddressForm();
+    },
+    editAddress(id) {
+      const address = this.addresses.find(address => address.id === id);
       this[SHOW_CREATE_NEW_ADDRESS_FORM](false);
+      this[SET_PROFILE_ADDRESS](address);
+      this[START_EDIT_ADDRESS](true);
     },
-    toggleNewAddressFormCreation() {
-      this[SHOW_CREATE_NEW_ADDRESS_FORM](true);
+    async saveAddress() {
+      this.errors = validateForm([
+        {
+          name: "name",
+          value: this.profileAddress.name,
+          rules: ["required"],
+        },
+        {
+          name: "street",
+          value: this.profileAddress.street,
+          rules: ["required"],
+        },
+        {
+          name: "building",
+          value: this.profileAddress.building,
+          rules: ["required"],
+        },
+      ]);
+      if (this.errors.length > 0) {
+        return;
+      }
+
+      const newAddress = {
+        name: this.profileAddress.name,
+        userId: this.user.id,
+        street: this.profileAddress.street,
+        building: this.profileAddress.building,
+        flat: this.profileAddress.flat,
+        comment: this.profileAddress.comment,
+      };
+
+      if (this.startedEditAddress) {
+        await this.putAddress({
+          id: this.profileAddress.id,
+          address: newAddress,
+        });
+      } else {
+        await this.postAddress(newAddress);
+      }
+
+      await this.fetchAddresses();
+      this.closeEditAddressForm();
     },
-    ...mapMutations("address", [SHOW_CREATE_NEW_ADDRESS_FORM]),
+    ...mapMutations("address", [
+      SHOW_CREATE_NEW_ADDRESS_FORM,
+      START_EDIT_ADDRESS,
+      SET_PROFILE_ADDRESS,
+      RESET_PROFILE_ADDRESS,
+    ]),
+    ...mapActions("address", [
+      "postAddress",
+      "putAddress",
+      "fetchAddresses",
+      "deleteAddress",
+    ]),
   },
 };
 </script>
